BREAKING CHANGES PLEASE READ - Deleted User folder and moved the files to Tenant folder - Created a Common folder to host files used by any script and not attached to any specifc feature. - Created a show package to bypass console.log - To use follow this example ->  const show = require(Common/Show.js);
