+++ src/Common/Show.js
@@ -1,0 +1,3 @@
+module.exports = show = (...args) => {
+  console.log(args);
+};
